[Design] #137 - BanchanDetail ui
